New wxWizardSample sample in Examples folder
The wxWizardSample demonstrates how to implment a wizard style interface
in wxphp. The only missing piece is the absence of a wxScrolledWindow
panel, which could be used to permit control panels that are larger than
the main wizard panel control space. This sample demonstrates this
limitation by gradually increasing the number of controls in the wizard
control space as the user continues to press Next.